requirements.txt
wget==3.2
python-nmap==0.6.1